+++ sdk/identity/identity/src/index.ts
@@ -26,7 +26,10 @@ export {
 } from "./credentials/environmentCredential";
 export { ClientSecretCredential } from "./credentials/clientSecretCredential";
 export { ClientSecretCredentialOptions } from "./credentials/clientSecretCredentialOptions";
-export { ClientCertificateCredential } from "./credentials/clientCertificateCredential";
+export {
+  ClientCertificateCredential,
+  ClientCertificateCredentialPEMConfiguration
+} from "./credentials/clientCertificateCredential";
 export { ClientCertificateCredentialOptions } from "./credentials/clientCertificateCredentialOptions";
 export { CredentialPersistenceOptions } from "./credentials/credentialPersistenceOptions";
 export { AzureCliCredential } from "./credentials/azureCliCredential";